feat(auth): implement and export configurePassport function for Passport.js setup
